Move downcast conversion helpers for selection to downcasthelpers.js.
